Generate stimuli a bit less randomly
Using normal distribution for weights and a bit tighter constraints
for bias/scale, we get results that are distributed a bit more similarly
to a "real" distribution. In the future, we need to make this controllable
from the outside so that we can perform some tests with these base settings
and others with more targeted ones.