Fix allowed types for neureka_v2
Neureka_v2 uses signed 8-bit scales. It should also support signed
32-bit scales but it is not tested.
Also for the output type, I have removed the 32-bit unquantized output
for now because there were some issues with the spatial strides (height
and width). There was the issue that output stride in the previous
neureka was multiplied with the number of output bytes, but in
neureka_v2 it seems that gets done in the accelerator, at least by the
look of the model. But when I tried with that change, the whole thing
broke. The output offsets were correct but still got all the wrong
results so I guess the output strides affect some other parts of the
model which I am not aware of.